🏗️ model connection과 scheme분리
