+++ src/com/team15/erp/model/customer/CustomerScheme.java
@@ -1,0 +1,37 @@
+package com.team15.erp.model.customer;
+
+public class CustomerScheme {
+    private String id;
+    private String name;
+    private String address;
+
+    public CustomerScheme(final String id, final String name, final String address) {
+        this.id = id;
+        this.name = name;
+        this.address = address;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(final String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(final String address) {
+        this.address = address;
+    }
+}
